pyproject.toml
[project]
name = "pythonproject"
version = "0.1.0"
description = "The L Project"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.187.0",
    "google-auth-oauthlib>=1.2.3",
    "matplotlib>=3.10.8",
]
